forcing: add NCOM ocean model interpolation routines
